✨(project) switch to a client/server architecture
We now run an HTTP server as the core Onzr server and interact with the
server using an HTTP client and a CLI using this client.

We now stream tracks over HTTP to VLC.

New commands were added to the CLI.

Note that we now use uv instead of poetry.